move retrieve_user_by_email function from graphql to account app, correct changelog
