+++ saleor/account/utils.py
@@ -112,3 +112,21 @@ def remove_staff_member(staff):
         staff.save()
     else:
         staff.delete()
+
+
+def retrieve_user_by_email(email):
+    """Retrieve user by email.
+
+    Email lookup is case-insensitive, unless the query returns more than one user. In
+    such a case, function return case-sensitive result.
+    """
+    users = User.objects.filter(email__iexact=email)
+
+    if len(users) > 1:
+        users_exact = [user for user in users if user.email == email]
+        users_iexact = [user for user in users if user.email == email.lower()]
+        users = users_exact or users_iexact  # type: ignore
+
+    if users:
+        return users[0]
+    return None
