+++ saleor/graphql/account/tests/mutations/test_token_create.py
@@ -1,13 +1,10 @@
 from datetime import datetime, timedelta, timezone
-from typing import List
 
 import graphene
-import pytest
 from django.urls import reverse
 from freezegun import freeze_time
 
 from .....account.error_codes import AccountErrorCode
-from .....account.models import User
 from .....core.jwt import (
     JWT_ACCESS_TYPE,
     JWT_REFRESH_TYPE,
@@ -16,7 +13,7 @@ from .....core.jwt import (
 )
 from .....core.utils import build_absolute_uri
 from ....tests.utils import get_graphql_content
-from ...mutations.authentication import CreateToken, _get_new_csrf_token
+from ...mutations.authentication import _get_new_csrf_token
 
 MUTATION_CREATE_TOKEN = """
     mutation tokenCreate($email: String!, $password: String!, $audience: String){
@@ -226,40 +223,20 @@ def test_create_token_active_user_logged_before(api_client, customer_user, setti
     assert payload["iss"] == build_absolute_uri(reverse("api"))
 
 
-@pytest.fixture
-def users_with_similar_emails():
-    users = [
-        User.objects.create_user("andrew@example.com", "password"),
-        User.objects.create_user("Andrew@example.com", "password"),
-        User.objects.create_user("john@example.com", "password"),
-        User.objects.create_user("Susan@example.com", "password"),
-        User.objects.create_user("Cindy@example.com", "password"),
-        User.objects.create_user("CINDY@example.com", "password"),
-    ]
-    return users
-
-
 @freeze_time("2020-03-18 12:00:00")
-@pytest.mark.parametrize(
-    "email,expected_user",
-    [
-        ("andrew@example.com", 0),
-        ("Andrew@example.com", 1),
-        ("ANDREW@example.com", 0),
-        ("john@example.com", 2),
-        ("John@example.com", 2),
-        ("Susan@example.com", 3),
-        ("susan@example.com", 3),
-        ("Cindy@example.com", 4),
-        ("cindy@example.com", None),
-        ("CiNdY@example.com", None),
-        ("non_existing_email@example.com", None),
-    ],
-)
-def test_email_case_sensitivity(email, expected_user, users_with_similar_emails):
+def test_create_token_email_case_insensitive(api_client, customer_user, settings):
     # given
-    users: List[User] = users_with_similar_emails
+    variables = {
+        "email": customer_user.email.upper(),
+        "password": customer_user._password,
+    }
+
     # when
-    user = CreateToken._retrieve_user_from_credentials(email=email, password="password")
+    response = api_client.post_graphql(MUTATION_CREATE_TOKEN, variables)
+    content = get_graphql_content(response)
+
     # then
-    assert user == users[expected_user] if expected_user is not None else user is None
+    data = content["data"]["tokenCreate"]
+    assert customer_user.email == data["user"]["email"]
+    assert not data["errors"]
+    assert data["token"]
